fix: navigate 수정
